Add files via upload
This Computer Vision based project is basically a text scanner that can read text from images.
The tkinter library was used to be able to select the required image from the local system, making the project more dynamic and user-friendly. The libraries used for the project were OpenCV,Numpy and Pytesseract. The images  should have the text located in a document and the document should have a colour contrast with its background in order to work well, the program can be adapted for different settings as well and be made more versatile but currently they are listed for future scope.